requirements.txt
discord.py==1.0.1
achallonge==1.8
tinydb==3.12.2